move hound stuff
