+++ src/lib.rs
@@ -5,6 +5,7 @@ pub mod dummy_hdm;
 pub mod hardware_data_manager;
 pub mod hardware_message_decoder;
 pub mod hdm;
+pub mod hound_helpers;
 pub mod localizer;
 pub mod saf;
 mod saf_raw;
